Fixed MonthlyStatistic

+++ src/main/java/com/bitmovin/bitcodin/api/statistics/MonthlyStatistic.java
@@ -6,7 +6,7 @@ public class MonthlyStatistic {
     @Expose
     public int jobCountFinished;
     @Expose
-    public int realtimeFactor;
+    public double realtimeFactor;
     @Expose
     public String totalBytesWritten;
     @Expose
